requirements.txt

scikit-learn
fastapi
uvicorn
ultralytics
joblib 
numpy 
pillow
opencv-python
pydantic
torch 
torchvision  
torchaudio   
streamlit
requests
python-multipart  